Added arduino sketch to manually deploy parachute
Work in progress. Capable of turning on a LED by sending a one char to
the xbee. Trying to make it work with a string. Most likely will only
need to use one arduino as we can send all the commands to the rocket
via a laptop hooked up to an xbee in transparent mode.